feat(Todo): Added bootsrap formatting

+++ Week08/day2/todo/src/main/java/com/greenfox/todo/controllers/TodoController.java
@@ -38,7 +38,7 @@ public class TodoController {
 
     @GetMapping("/{id}/delete")
     public String delete(@PathVariable Long id) {
-        todoRepository.deleteById(id);
+        if (todoRepository.findById(id).isPresent()) todoRepository.deleteById(id);
         return "redirect:/todo";
     }
